:construction: feat: showing CK activities
